// frontend/ticket-booking-web/src/app/api/model/auth-result-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResultDto { 
    accessToken: string;
    expiresAt: string;
}